requirements.txt
gradio
easyocr
torch
torchvision
opencv-python-headless
python-docx
numpy